Sync protocol structs and Tauri SPA build
Refactored the protocol client/ops layer to match canonical memdbg_protocol.h wire layouts across hello, debugger, tracer, task manager, memory, quickscan, ptwalk, and admin paths, updated dependent session/store call sites, and added end-to-end protocol tests with a mock transport for framing, status, and timeout behavior. Switched the frontend to a Tauri-friendly SPA entrypoint (new static index/main.tsx, router root cleanup, and Vite plugin/build changes), added post-build HTML generation for hashed assets, and introduced a macOS GitHub Actions workflow that builds and uploads .dmg/.app artifacts.

+++ memdbg-tauri/src/components/workspace/DebuggerPanel.tsx
@@ -259,10 +259,9 @@ function RegRow({
       <input
         value={text}
         onChange={(e) => setText(e.target.value)}
-        onBlur={() => {
-          try {
-            const v = hexToAddr(text);
-            if (v !== value) onCommit(v);
+        onBlur={() => {            try {
+              const v = hexToAddr(text);
+              if (v !== value) onCommit(v as bigint);
           } catch {
             setText(addrToHex(value));
           }
